requirements.txt
aioredis[hiredis]==2.0.1
fastapi==0.75.2
pydantic==1.9.0
# for fixtures
pyyaml==6.0
sentry-sdk==1.5.10
ujson==5.2.0
uvicorn[standard]==0.17.6
